Separate client and server interceptor tests
