+++ src/test/java/io/opentracing/contrib/grpc/ClientTracingInterceptorTest.java
@@ -1,0 +1,317 @@
+/*
+ * Copyright 2017-2019 The OpenTracing Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package io.opentracing.contrib.grpc;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+
+import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.ImmutableSet;
+import io.grpc.CallOptions;
+import io.grpc.Metadata;
+import io.grpc.MethodDescriptor;
+import io.grpc.Status;
+import io.grpc.testing.GrpcServerRule;
+import io.opentracing.Span;
+import io.opentracing.log.Fields;
+import io.opentracing.mock.MockSpan;
+import io.opentracing.mock.MockTracer;
+import io.opentracing.tag.Tags;
+import io.opentracing.util.GlobalTracerTestUtil;
+import org.assertj.core.api.Assertions;
+import org.assertj.core.data.MapEntry;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+public class ClientTracingInterceptorTest {
+
+  private static final Map<String, Object> BASE_TAGS = ImmutableMap.<String, Object>builder()
+      .put(Tags.COMPONENT.getKey(), GrpcTags.COMPONENT_NAME)
+      .put(GrpcTags.GRPC_STATUS.getKey(), Status.Code.OK.name())
+      .build();
+
+  private static final Map<String, Object> BASE_CLIENT_TAGS = ImmutableMap.<String, Object>builder()
+      .putAll(BASE_TAGS)
+      .put(Tags.SPAN_KIND.getKey(), Tags.SPAN_KIND_CLIENT)
+      .build();
+
+  private static final Set<String> CLIENT_ATTRIBUTE_TAGS = ImmutableSet.of(
+      GrpcTags.GRPC_CALL_OPTIONS.getKey(),
+      //TODO: @ExperimentalApi("https://github.com/grpc/grpc-java/issues/1767")
+      //GrpcTags.GRPC_AUTHORITY.getKey(),
+      GrpcTags.GRPC_COMPRESSOR.getKey(),
+      GrpcTags.GRPC_DEADLINE.getKey(),
+      GrpcTags.GRPC_METHOD_NAME.getKey(),
+      GrpcTags.GRPC_METHOD_TYPE.getKey(),
+      GrpcTags.GRPC_HEADERS.getKey());
+
+  private final MockTracer clientTracer = new MockTracer();
+  private final MockTracer serverTracer = new MockTracer();
+
+  @Rule
+  public GrpcServerRule grpcServer = new GrpcServerRule();
+
+  private TracedService service;
+
+  @Before
+  public void before() {
+    GlobalTracerTestUtil.resetGlobalTracer();
+    clientTracer.reset();
+    serverTracer.reset();
+    service = new TracedService();
+  }
+
+  @Test
+  public void testTracedClientBasic() {
+    ClientTracingInterceptor tracingInterceptor = new ClientTracingInterceptor(clientTracer);
+    TracedClient client = new TracedClient(grpcServer.getChannel(), tracingInterceptor);
+
+    service.addGreeterService(grpcServer.getServiceRegistry());
+
+    assertTrue("call should complete", client.greet("world"));
+    assertEquals("one span should have been created and finished for one client request",
+        clientTracer.finishedSpans().size(), 1);
+
+    MockSpan span = clientTracer.finishedSpans().get(0);
+    assertEquals("span should have prefix", span.operationName(), "helloworld.Greeter/SayHello");
+    assertEquals("span should have no parents", span.parentId(), 0);
+    assertEquals("span should have no logs", span.logEntries().size(), 0);
+    Assertions.assertThat(span.tags()).as("span should have base client tags")
+        .isEqualTo(BASE_CLIENT_TAGS);
+    assertFalse("span should have no baggage",
+        span.context().baggageItems().iterator().hasNext());
+  }
+
+  @Test
+  public void testTracedClientTwoInterceptors() {
+    SecondClientInterceptor secondClientInterceptor = new SecondClientInterceptor(clientTracer);
+    ClientTracingInterceptor clientTracingInterceptor = new ClientTracingInterceptor(clientTracer);
+    TracedClient client = new TracedClient(grpcServer.getChannel(), secondClientInterceptor, clientTracingInterceptor);
+
+    service.addGreeterService(grpcServer.getServiceRegistry());
+
+    assertTrue("call should complete", client.greet("world"));
+    assertEquals("one span should have been created and finished for one client request",
+        clientTracer.finishedSpans().size(), 1);
+
+    MockSpan span = clientTracer.finishedSpans().get(0);
+    assertEquals("span should have prefix", span.operationName(), "helloworld.Greeter/SayHello");
+    assertEquals("span should have no parents", span.parentId(), 0);
+    assertEquals("span should have no logs", span.logEntries().size(), 0);
+    Assertions.assertThat(span.tags()).as("span should have base client tags")
+        .isEqualTo(BASE_CLIENT_TAGS);
+    assertFalse("span should have no baggage",
+        span.context().baggageItems().iterator().hasNext());
+  }
+
+  @Test
+  public void testTracedClientWithVerbosity() {
+    ClientTracingInterceptor tracingInterceptor = new ClientTracingInterceptor
+        .Builder(clientTracer)
+        .withVerbosity()
+        .build();
+    TracedClient client = new TracedClient(grpcServer.getChannel(), tracingInterceptor);
+
+    service.addGreeterService(grpcServer.getServiceRegistry());
+
+    assertTrue("call should complete", client.greet("world"));
+    assertEquals("one span should have been created and finished for one client request",
+        clientTracer.finishedSpans().size(), 1);
+
+    MockSpan span = clientTracer.finishedSpans().get(0);
+    assertEquals("span should have prefix", span.operationName(), "helloworld.Greeter/SayHello");
+    assertEquals("span should have no parents", span.parentId(), 0);
+    System.out.println(span.logEntries());
+    List<String> events = new ArrayList<>(span.logEntries().size());
+    for (MockSpan.LogEntry logEntry : span.logEntries()) {
+      events.add((String) logEntry.fields().get(Fields.EVENT));
+    }
+    Assertions.assertThat(events)
+        .as("span should contain verbose log fields")
+        .contains(
+            GrpcFields.CLIENT_CALL_START,
+            GrpcFields.CLIENT_CALL_SEND_MESSAGE,
+            GrpcFields.CLIENT_CALL_HALF_CLOSE,
+            GrpcFields.CLIENT_CALL_LISTENER_ON_HEADERS,
+            GrpcFields.CLIENT_CALL_LISTENER_ON_MESSAGE,
+            GrpcFields.CLIENT_CALL_LISTENER_ON_CLOSE);
+    Assertions.assertThat(span.tags()).as("span should have base client tags")
+        .isEqualTo(BASE_CLIENT_TAGS);
+    assertFalse("span should have no baggage",
+        span.context().baggageItems().iterator().hasNext());
+  }
+
+  @Test
+  public void testTracedClientWithStreaming() {
+    ClientTracingInterceptor tracingInterceptor = new ClientTracingInterceptor
+        .Builder(clientTracer)
+        .withStreaming()
+        .build();
+    TracedClient client = new TracedClient(grpcServer.getChannel(), tracingInterceptor);
+
+    service.addGreeterService(grpcServer.getServiceRegistry());
+
+    assertTrue("call should complete", client.greet("world"));
+    assertEquals("one span should have been created and finished for one client request",
+        clientTracer.finishedSpans().size(), 1);
+
+    MockSpan span = clientTracer.finishedSpans().get(0);
+    assertEquals("span should have prefix", span.operationName(), "helloworld.Greeter/SayHello");
+    assertEquals("span should have no parents", span.parentId(), 0);
+    List<String> events = new ArrayList<>(span.logEntries().size());
+    for (MockSpan.LogEntry logEntry : span.logEntries()) {
+      events.add((String) logEntry.fields().get(Fields.EVENT));
+    }
+    Assertions.assertThat(events)
+        .as("span should contain streaming log fields")
+        .contains(
+            GrpcFields.CLIENT_CALL_SEND_MESSAGE,
+            GrpcFields.CLIENT_CALL_HALF_CLOSE,
+            GrpcFields.CLIENT_CALL_LISTENER_ON_MESSAGE);
+    Assertions.assertThat(span.tags()).as("span should have base client tags")
+        .isEqualTo(BASE_CLIENT_TAGS);
+    assertFalse("span should have no baggage",
+        span.context().baggageItems().iterator().hasNext());
+  }
+
+  @Test
+  public void testTracedClientWithOperationName() {
+    final String PREFIX = "testing-";
+
+    ClientTracingInterceptor tracingInterceptor = new ClientTracingInterceptor
+        .Builder(clientTracer)
+        .withOperationName(new OperationNameConstructor() {
+          @Override
+          public <ReqT, RespT> String constructOperationName(MethodDescriptor<ReqT, RespT> method) {
+            return PREFIX + method.getFullMethodName();
+          }
+        })
+        .build();
+    TracedClient client = new TracedClient(grpcServer.getChannel(), tracingInterceptor);
+
+    service.addGreeterService(grpcServer.getServiceRegistry());
+
+    assertTrue("call should complete", client.greet("world"));
+    assertEquals("one span should have been created and finished for one client request",
+        clientTracer.finishedSpans().size(), 1);
+
+    MockSpan span = clientTracer.finishedSpans().get(0);
+    assertEquals("span should have prefix", span.operationName(),
+        PREFIX + "helloworld.Greeter/SayHello");
+    assertEquals("span should have no parents", span.parentId(), 0);
+    assertEquals("span should have no logs", span.logEntries().size(), 0);
+    Assertions.assertThat(span.tags()).as("span should have base client tags")
+        .isEqualTo(BASE_CLIENT_TAGS);
+    assertFalse("span should have no baggage",
+        span.context().baggageItems().iterator().hasNext());
+  }
+
+  @Test
+  public void testTracedClientWithTracedAttributes() {
+    ClientTracingInterceptor tracingInterceptor = new ClientTracingInterceptor
+        .Builder(clientTracer)
+        .withTracedAttributes(ClientTracingInterceptor.ClientRequestAttribute.values())
+        .build();
+    TracedClient client = new TracedClient(grpcServer.getChannel(), tracingInterceptor);
+
+    service.addGreeterService(grpcServer.getServiceRegistry());
+
+    assertTrue("call should complete", client.greet("world"));
+    assertEquals("one span should have been created and finished for one client request",
+        clientTracer.finishedSpans().size(), 1);
+
+    MockSpan span = clientTracer.finishedSpans().get(0);
+    assertEquals("span should have prefix", span.operationName(), "helloworld.Greeter/SayHello");
+    assertEquals("span should have no parents", span.parentId(), 0);
+    assertEquals("span should have no logs", span.logEntries().size(), 0);
+    Assertions.assertThat(span.tags()).as("span should have base client tags")
+        .containsAllEntriesOf(BASE_CLIENT_TAGS);
+    Assertions.assertThat(span.tags().keySet())
+        .as("span should have tags for all client request attributes")
+        .containsAll(CLIENT_ATTRIBUTE_TAGS);
+    assertFalse("span should have no baggage",
+        span.context().baggageItems().iterator().hasNext());
+  }
+
+  @Test
+  public void testTracedClientWithClientSpanDecorator() {
+    ClientSpanDecorator clientSpanDecorator = new ClientSpanDecorator() {
+      @Override
+      public void interceptCall(Span span, MethodDescriptor method, CallOptions callOptions) {
+        span.setTag("test_tag", "test_value");
+        span.setTag("tag_from_method", method.getFullMethodName());
+        span.setTag("tag_from_call_options", callOptions.getCompressor());
+
+        span.log("A test log");
+      }
+    };
+
+    ClientTracingInterceptor tracingInterceptor = new ClientTracingInterceptor
+        .Builder(clientTracer)
+        .withClientSpanDecorator(clientSpanDecorator)
+        .build();
+    TracedClient client = new TracedClient(grpcServer.getChannel(), tracingInterceptor);
+
+    service.addGreeterService(grpcServer.getServiceRegistry());
+
+    assertTrue("call should complete", client.greet("world"));
+    assertEquals("one span should have been created and finished for one client request",
+        clientTracer.finishedSpans().size(), 1);
+
+    MockSpan span = clientTracer.finishedSpans().get(0);
+    assertEquals("span should have prefix", span.operationName(), "helloworld.Greeter/SayHello");
+    assertEquals("span should have no parents", span.parentId(), 0);
+    assertEquals("span should have one log from the decorator",
+        span.logEntries().size(), 1);
+    Assertions.assertThat(span.tags()).as("span should have 3 tags from the decorator")
+        .hasSize(3 + BASE_CLIENT_TAGS.size());
+    Assertions.assertThat(span.tags()).as("span contains added tags")
+        .contains(MapEntry.entry("test_tag", "test_value"))
+        .containsKeys("tag_from_method", "tag_from_call_options");
+    assertFalse("span should have no baggage",
+        span.context().baggageItems().iterator().hasNext());
+  }
+
+  @Test
+  public void testTracedClientWithClientCloseDecorator() {
+    ClientCloseDecorator clientCloseDecorator = new ClientCloseDecorator() {
+      @Override
+      public void close(Span span, Status status, Metadata trailers) {
+        span.setTag("grpc.statusCode", status.getCode().value());
+      }
+    };
+
+    ClientTracingInterceptor tracingInterceptor = new ClientTracingInterceptor
+        .Builder(clientTracer)
+        .withClientCloseDecorator(clientCloseDecorator)
+        .build();
+    TracedClient client = new TracedClient(grpcServer.getChannel(), tracingInterceptor);
+
+    service.addGreeterService(grpcServer.getServiceRegistry());
+
+    assertTrue("call should complete", client.greet("world"));
+    assertEquals("one span should have been created and finished for one client request",
+        clientTracer.finishedSpans().size(), 1);
+
+    MockSpan span = clientTracer.finishedSpans().get(0);
+    Assertions.assertThat(span.tags())
+        .contains(MapEntry.entry("grpc.statusCode", Status.OK.getCode().value()));
+  }
+}
